break out different sub-scenarios into their own functions and store sub-scenarios into it's own file
